requirements.txt
faicons
joblib
pandas
plotly
python_dateutil
Requests
shiny
shinywidgets
sqlitecloud
openpyxl